+++ tcs/powerblock.cpp
@@ -1289,7 +1289,7 @@ endif
 
 	//Use brute force interpolation.. it is faster in this case than bisection or hunting methods used in the user-specified HTF case
 	
-	int iLastIndex = m_db.ncols()-1;
+	int iLastIndex = (int)m_db.ncols() - 1;
 	for (size_t i=0; i < m_db.ncols(); i++)
 	{
 		// if we got to the last one, then set bounds and end loop
@@ -1330,8 +1330,8 @@ endif
 		// so the reference [i+1], where i = iLastIndex, will never happen
 		if( ( (X >= m_db.at(XI,i)) && (X < m_db.at(XI,i+1)) ) || ( (X <= m_db.at(XI,i)) && (X > m_db.at(XI,i+1)) ) )
 		{
-			lbi = i;
-			ubi = i+1;
+			lbi = (int)i;
+			ubi = (int)i + 1;
 			break;
 		}
 	}
